Send data from mixer to APM limiter more often.
Before this change, the APM limiter used in FrameCombiner (a
sub-component of AudioMixer) only gets to process the data when the
number of non-muted streams is >1. If this number varies between <=1
and >1, the limiter's view of the data will have gaps during the
periods with <= 1 active stream.

This leads to discontinuities in the applied gain. These
discontinuities cause clicks in the output audio. This change
activates APM limiter processing based on the number of audio streams,
independently of their mutedness status.

BUG=chromium:695993
TBR=aleloi@webrtc.org

Review-Url: https://codereview.webrtc.org/2776113002
Cr-Original-Commit-Position: refs/heads/master@{#17442}
Review-Url: https://codereview.webrtc.org/2788683002 .
Cr-Commit-Position: refs/branch-heads/58@{#11}
Cr-Branched-From: f31969a584bcafe9406c214a9d4c3afb49d19650-refs/heads/master@{#16937}